refactor(TDKN-140):moved migration interfaces to a migration package
