+++ daikon/src/main/java/org/talend/daikon/serialize/PostDeserializeSetup.java
@@ -1,5 +1,7 @@
 package org.talend.daikon.serialize;
 
+import org.talend.daikon.serialize.migration.PostDeserializeHandler;
+
 /**
  * Interface that allows code to be injected into the post-deserialization processing.
  */
