+++ daikon/src/test/java/org/talend/daikon/serialize/jsonio/PersistenceTestObjectInner2.java
@@ -1,10 +1,10 @@
 package org.talend.daikon.serialize.jsonio;
 
-import org.talend.daikon.serialize.DeserializeDeletedFieldHandler;
 import org.talend.daikon.serialize.PersistenceTestObjectInner3;
-import org.talend.daikon.serialize.PostDeserializeHandler;
 import org.talend.daikon.serialize.PostDeserializeSetup;
-import org.talend.daikon.serialize.SerializeSetVersion;
+import org.talend.daikon.serialize.migration.DeserializeDeletedFieldHandler;
+import org.talend.daikon.serialize.migration.PostDeserializeHandler;
+import org.talend.daikon.serialize.migration.SerializeSetVersion;
 
 public class PersistenceTestObjectInner2 implements DeserializeDeletedFieldHandler, PostDeserializeHandler, SerializeSetVersion {
